requirements.txt
gradio_imageslider
gradio==4.36.0
torchmetrics
tornado==6.5.2
opencv-python
matplotlib
huggingface_hub
einops
safetensors
open3d
tensorboardX
mmengine
pyexr
